requirements.txt
pylint
flask
gunicorn
google-cloud-secret-manager